pyproject.toml
[tool.poetry]
name = "shopnext"
version = "0.1.0"
description = ""
authors = ["Ankush <ankush357159@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django-cors-headers = "^3.10.1"
djangorestframework = "^3.13.1"
Pillow = "^9.0.0"
djangorestframework-simplejwt = "^5.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
